Removed unused dependencies. Added figment. Refactoring continued. Adding tracing & configs.
