+++ Cargo.toml
@@ -4,22 +4,18 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-axum = { version = "0.7.2", features = ["http2"] }
+axum = { version = "0.7", features = ["http2"] }
 bip32 = "0.5"
 candid = "0.10"
 chrono = "0.4"
-der = "0.7"
-hex = "0.4.3"
-http = "1.0"
-ic-agent = "0.31.0"
+figment = { version = "0.10", features = ["env", "toml"] }
+ic-agent = "0.31"
 k256 = "0.13"
-rand = "0.8"
 sec1 = "0.7"
 serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
 thiserror = "1.0"
 tiny-bip39 = "1.0"
-tokio = { version = "1.35.1", features = ["rt-multi-thread", "macros"] }
+tokio = { version = "1.35", features = ["rt-multi-thread", "macros"] }
 tower = "0.4"
 tower-http = { version = "0.5", features = ["cors"] }
 tracing = "0.1"
