+++ examples/verify-auth-leptos/src/app.rs
@@ -9,8 +9,6 @@ pub fn App() -> impl IntoView {
     provide_meta_context();
 
     view! {
-
-
         // injects a stylesheet into the document <head>
         // id=leptos means cargo-leptos will hot-reload this stylesheet
         <Stylesheet id="leptos" href="/pkg/verify-auth-leptos.css"/>
